program now represented as a byte array

+++ msim/programloader.hpp
@@ -1,6 +1,8 @@
 #ifndef PROGRAMLOADER_H
 #define PROGRAMLOADER_H
 
-int load_program(char *program_name, int **program);
+#include <cstdint>
+
+int load_program(char *program_name, uint8_t **program);
 
 #endif
